+++ presentation/src/main/java/com/threegap/bitnagil/presentation/routinelist/RoutineListViewModel.kt
@@ -9,8 +9,8 @@ import com.threegap.bitnagil.domain.routine.usecase.DeleteRoutineUseCase
 import com.threegap.bitnagil.domain.routine.usecase.FetchWeeklyRoutinesUseCase
 import com.threegap.bitnagil.domain.writeroutine.usecase.GetWriteRoutineEventFlowUseCase
 import com.threegap.bitnagil.presentation.home.util.getCurrentWeekDays
-import com.threegap.bitnagil.presentation.routinelist.model.RoutineListSideEffect
-import com.threegap.bitnagil.presentation.routinelist.model.RoutineListState
+import com.threegap.bitnagil.presentation.routinelist.contract.RoutineListSideEffect
+import com.threegap.bitnagil.presentation.routinelist.contract.RoutineListState
 import com.threegap.bitnagil.presentation.routinelist.model.RoutineUiModel
 import com.threegap.bitnagil.presentation.routinelist.model.toUiModel
 import dagger.hilt.android.lifecycle.HiltViewModel
